Combine unittest module files into one
